Added progress view (new to ios14)

+++ Agendum/Models/User.swift
@@ -15,5 +15,5 @@ struct User {
     var uid: String
     var items: Array<Item>
     var labels: Array<String>
-    var progress: Int
+    var progress: Double
 }
